+++ 1.21.8/CHANGELOG.md
@@ -1,8 +1,27 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.8.1-1.21.8] - 2025-09-16
+
+### Added
+
+- Add a config option for doubling the player base attack damage (disabled by default)
+
+### Changed
+
+- Allow defining a custom `minHitboxSize` value
+- Separately support toggling eating and drinking for the `eatingInterruption` option
+
+### Fixed
+
+- Fix item attributes with an amount of zero being shown on tooltips (like armour toughness)
+
 ## [v21.8.0-1.21.8] - 2025-07-18
+
+### Changed
+
 - Update to Minecraft 1.21.8